refactor: update gadget handling and improve logging; remove unused Association component
